Fix broken links, slow loading, and prod errors
