+++ packages/twenty-website/src/app/_components/docs/DocsContent.tsx
@@ -1,7 +1,6 @@
 'use client';
 import React, { useState } from 'react';
 import styled from '@emotion/styled';
-import Image from 'next/image';
 import { usePathname } from 'next/navigation';
 
 import { ArticleContent } from '@/app/_components/ui/layout/articles/ArticleContent';
@@ -112,7 +111,7 @@ const StyledImageContainer = styled.div<{ loaded: string }>`
   `}
 `;
 
-const StyledImage = styled(Image)<{ loaded: string }>`
+const StyledImage = styled(img)<{ loaded: string }>`
   opacity: ${({ loaded }) => (loaded === 'true' ? 1 : 0)};
   transition: opacity 250ms ease-in-out;
 `;
