+++ packages/twenty-website/src/app/developers/rest-api/metadata/page.tsx
@@ -1,39 +1,22 @@
 'use client';
 import React, { useEffect, useState } from 'react';
-// @ts-expect-error Migration loader as text not passing warnings
-import { API } from '@stoplight/elements';
 
 import Playground from '@/app/_components/playground/playground';
-
-// @ts-expect-error Migration loader as text not passing warnings
-import spotlightTheme from '!css-loader!@stoplight/elements/styles.min.css';
-
-const RestApiComponent = ({ openApiJson }: { openApiJson: any }) => {
-  // We load spotlightTheme style using useEffect as it breaks remaining docs style
-  useEffect(() => {
-    const styleElement = document.createElement('style');
-    styleElement.innerHTML = spotlightTheme.toString();
-    document.head.append(styleElement);
-
-    return () => styleElement.remove();
-  }, []);
-
-  return (
-    <div
-      style={{
-        height: 'calc(100vh - var(--ifm-navbar-height) - 45px)',
-        overflow: 'auto',
-      }}
-    >
-      <API apiDescriptionDocument={JSON.stringify(openApiJson)} router="hash" />
-    </div>
-  );
-};
+import { RestApiWrapper } from '@/app/_components/playground/rest-api-wrapper';
 
 const restApi = () => {
   const [openApiJson, setOpenApiJson] = useState({});
+  const [isClient, setIsClient] = useState(false);
 
-  const children = <RestApiComponent openApiJson={openApiJson} />;
+  useEffect(() => {
+    setIsClient(true);
+  }, []);
+
+  if (!isClient) {
+    return null;
+  }
+
+  const children = <RestApiWrapper openApiJson={openApiJson} />;
 
   return (
     <Playground
